requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==1.10.8
django-braces==1.12.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-formtools==1.0
django-guardian==1.4.9
django-haystack==2.7.0
django-import-export==1.0.0
django-ipware==2.0.1
django-markdown-deux==1.0.5
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.1.0
django-reversion==2.0.13
django-simple-captcha==0.4.6
djangorestframework==3.7.7
drf-extensions==0.3.1
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.6
itypes==1.1.0
jdcal==1.4
jieba==0.39
Jinja2==2.10
Markdown==2.6.11
markdown2==2.3.5
MarkupSafe==1.0
MySQL-python==1.2.5
oauthlib==2.0.7
odfpy==1.3.6
olefile==0.44
openpyxl==2.5.3
Pillow==4.3.0
Pygments==2.2.0
PyJWT==1.6.1
python-memcached==1.58
python-openid==2.2.5
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
